@@ -1,39 +1,40 @@
 {
   "name": "alienlayout",
   "description": "Twoday Alien Template to host a foreign Wordpress site",
-  "version": "1.10.2",
+  "version": "1.11.0",
   "author": "NeonWilderness",
   "bugs": {
     "url": "https://github.com/NeonWilderness/tdalien/issues"
   },
   "devDependencies": {
-    "@babel/core": "^7.16.0",
-    "@babel/preset-env": "^7.16.0",
-    "@neonwilderness/twoday": "^0.3.17",
-    "@types/jquery": "^3.5.8",
-    "babel-loader": "^8.2.3",
-    "css-loader": "^6.5.1",
+    "@babel/core": "^7.19.1",
+    "@babel/eslint-parser": "^7.19.1",
+    "@babel/preset-env": "^7.19.1",
+    "@neonwilderness/twoday": "^0.3.22",
+    "@types/jquery": "^3.5.14",
+    "@types/toastr": "^2.1.40",
+    "babel-loader": "^8.2.5",
+    "css-loader": "^6.7.1",
     "dotenv-safe": "^8.2.0",
-    "eslint": "^7.32.0",
-    "eslint-config-standard": "^16.0.3",
-    "eslint-friendly-formatter": "^4.0.1",
-    "eslint-loader": "^4.0.2",
-    "eslint-plugin-import": "^2.25.2",
+    "eslint": "^8.23.1",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^5.1.1",
+    "eslint-plugin-promise": "^6.0.1",
     "eslint-plugin-standard": "^5.0.0",
-    "html-loader": "^3.0.1",
+    "eslint-webpack-plugin": "^3.2.0",
+    "html-loader": "^4.1.0",
     "html-webpack-plugin": "^5.5.0",
-    "less": "^4.1.2",
-    "less-loader": "^10.2.0",
-    "mini-css-extract-plugin": "^2.4.4",
-    "node-fetch": "^3.0.0",
-    "semver": "^7.3.5",
-    "string-replace-loader": "^3.0.3",
-    "url-join": "^4.0.1",
-    "webpack": "^5.62.1",
-    "webpack-cli": "^4.9.1",
-    "yargs": "^17.2.1"
+    "less": "^4.1.3",
+    "less-loader": "^11.0.0",
+    "mini-css-extract-plugin": "^2.6.1",
+    "node-fetch": "^3.2.10",
+    "semver": "^7.3.7",
+    "string-replace-loader": "^3.1.0",
+    "url-join": "^5.0.0",
+    "webpack": "^5.74.0",
+    "webpack-cli": "^4.10.0",
+    "yargs": "^17.5.1"
   },
   "homepage": "https://github.com/NeonWilderness/tdalien/blob/master/README.md",
   "keywords": [
@@ -49,11 +50,13 @@
     "url": "git://github.com/NeonWilderness/tdalien.git"
   },
   "scripts": {
-    "update": "ncu -u -x eslint && npm i",
+    "update": "ncu -u && npm i",
     "build": "webpack --mode production",
     "deploy": "node ./utils/uploadSkin --alias=foundation --platform=prod",
-    "deploy:client": "node ./utils/uploadSkin --alias=xxx --platform=prod",
+    "deploy:client": "node ./utils/uploadSkin --alias=weltentanz --platform=prod",
     "dev": "webpack --mode development",
+    "lint": "eslint --ext .js ./src --no-error-on-unmatched-pattern",
+    "lint:utils": "eslint --ext .js ./utils --no-error-on-unmatched-pattern",
     "major": "node ./utils/bump --mode=major",
     "minor": "node ./utils/bump --mode=minor",
     "patch": "node ./utils/bump --mode=patch",
